requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
SQLAlchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0
pytest==8.1.1
Pillow==10.2.0
numpy==1.26.4 